Move methods to version specific and update links for validate
